# docs/swagger.yaml
definitions:
  fibererror.ResponseError:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  models.CreateUserRequest:
    properties:
      password:
        example: password123
        type: string
      user_name:
        example: john_doe
        type: string
    type: object
  models.PingPongResponse:
    properties:
      is_success:
        type: boolean
      message:
        type: string
      timestamp:
        type: string
    type: object
  models.UpdateUserPasswordRequest:
    properties:
      new_password:
        example: newpassword123
        type: string
    type: object
  models.UserResponse:
    properties:
      user_name:
        example: john_doe
        type: string
    type: object
  models.UsersListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.UserResponse'
        type: array
      message:
        example: Users retrieved successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/users:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all registered users
      parameters:
      - description: KodeDC
        in: query
        name: KodeDC
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            $ref: '#/definitions/models.UsersListResponse'
        "500":
          description: Failed to fetch users
          schema:
            $ref: '#/definitions/fibererror.ResponseError'
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account with username and password
      parameters:
      - description: User creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or missing fields
          schema:
            $ref: '#/definitions/fibererror.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fibererror.ResponseError'
      summary: Create a new user
      tags:
      - users
  /api/users/{user_name}:
    delete:
      consumes:
      - application/json
      description: Delete a user account by username
      parameters:
      - description: Username of the user to delete
        in: path
        name: user_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete user
          schema:
            $ref: '#/definitions/fibererror.ResponseError'
      summary: Delete a user
      tags:
      - users
  /api/users/{user_name}/password:
    put:
      consumes:
      - application/json
      description: Update the password for a specific user by username
      parameters:
      - description: Username of the user to update
        in: path
        name: user_name
        required: true
        type: string
      - description: New password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or missing fields
          schema:
            $ref: '#/definitions/fibererror.ResponseError'
        "500":
          description: Failed to update password
          schema:
            $ref: '#/definitions/fibererror.ResponseError'
      summary: Update user password
      tags:
      - users
  /ping:
    get:
      description: Ping Pong endpoint to check service availability
      produces:
      - application/json
      responses:
        "200":
          description: Pong response
          schema:
            $ref: '#/definitions/models.PingPongResponse'
      summary: Ping Check
      tags:
      - _
  /swagger:
    get:
      description: Interactive Swagger UI for API documentation
      produces:
      - text/html
      responses:
        "200":
          description: Swagger UI HTML
          schema:
            type: string
      summary: Swagger UI
      tags:
      - Swagger
  /swagger/doc.json:
    get:
      description: Returns the Swagger/OpenAPI documentation in JSON format with proxy
        path support
      produces:
      - application/json
      responses:
        "200":
          description: Swagger documentation
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Swagger API Documentation
      tags:
      - Swagger
swagger: "2.0"